move server code into directory
