+++ server/views/login.hbs
@@ -1,0 +1,22 @@
+<form method="POST" action="login">
+    {{#if message}}
+        <div id="divMessage">
+            {{message}}
+        </div>
+    {{/if}}
+
+    <div>
+        <input type="text" placeholder="Username" name="username">
+    </div>
+    <div>
+        <input type="password" placeholder="Password"  name="password">
+    </div>
+    <div>
+        <input type="submit" value="Log In">
+    </div>
+
+    <div>
+        Not a member yet? <a href="/register"> Sign Up </a>
+    </div>
+    <!-- <div><a href=""> Forgot Password? </a></div> -->
+</form>
