--- views/drink_view.hbs
@@ -1,22 +1,0 @@
-<H2>
-  {{#if isNew}}
-    New Drink
-  {{else}}
-    {{name}}
-  {{/if}}
-</H2>
-
-<form method="POST" action="/api/drinks">
-  Name:<br>
-  <input type="text" name="name"><br>
-  Ingredients:<br>
-  <input type="text" name="ingredients[]"><br>
-  <input type="text" name="ingredients[]"><br>
-  <input type="text" name="ingredients[]"><br>
-  <input type="text" name="ingredients[]"><br>
-  <input type="text" name="ingredients[]"><br>
-  Instructions:<br>
-  <textarea name="instructions">
-  </textarea><br>
-  <input type="submit" value="Create">
-</form>
