--- views/drinks.hbs
@@ -1,11 +1,0 @@
-<h2>
-  Drinks List
-</h2>
-<div>
-  <a href="/drinks/create">Create New Drink</a>
-</div>
-{{#each drinks}}
-  <div>
-    <a href="{{url}}">{{name}}</a>
-  </div> 
-{{/each}}
