--- views/register.hbs
@@ -1,26 +1,0 @@
-<form method="POST">
-    {{#if message}}
-        <div id="divMessage">
-            {{message}}
-        </div>
-    {{/if}}
-
-    <div>
-        <input type="text" placeholder="Username" name="username">
-    </div>
-    <div>
-        <input type="password" placeholder="Password"  name="password">
-    </div>
-    <div>
-        <input type="password" placeholder="Confirm Password" name="verify password">
-    </div>
-    <div>
-        <input type="text" placeholder="First Name" name="first_name">
-    </div>
-    <div>
-        <input type="text" placeholder="Last Name" name="last_name">
-    </div>
-    <div>
-        <input type="submit" value="Register">
-    </div>
-</form>
